Read and log bytes before writing response
I'm worried that something on the car's end will not be ready to read
until it has finished writing. So let's only write our response after
we've read what the car sent.

I'm also curious to know what it is that the car is sending, in case
this can help us be more resilient or responsive. So let's log what it
sends.